move lexer and parser to correct package
